requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
PyJWT==2.1.0
requests==2.26.0
urllib3==1.26.7
python-keycloak==0.27.0
cryptography==36.0.1
